ges: tests: Use correct variable types when setting properties
From https://gitlab.freedesktop.org/gstreamer/gst-editing-services/-/merge_requests/247

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/4995>

+++ subprojects/gst-editing-services/tests/check/nle/nlecomposition.c
@@ -228,7 +228,8 @@ GST_START_TEST (test_remove_last_object)
   audiotestsrc = gst_element_factory_make ("audiotestsrc", "audiotestsrc1");
   gst_bin_add (GST_BIN (source1), audiotestsrc);
   g_object_set (source1, "start", (guint64) 0 * GST_SECOND,
-      "duration", 10 * GST_SECOND, "inpoint", (guint64) 0, "priority", 1, NULL);
+      "duration", (gint64) 10 * GST_SECOND, "inpoint", (guint64) 0, "priority",
+      1, NULL);
 
   nle_composition_add (composition, source1);
 
@@ -236,7 +237,8 @@ GST_START_TEST (test_remove_last_object)
   audiotestsrc2 = gst_element_factory_make ("audiotestsrc", "audiotestsrc1");
   gst_bin_add (GST_BIN (source2), audiotestsrc2);
   g_object_set (source2, "start", (guint64) 10 * GST_SECOND,
-      "duration", 10 * GST_SECOND, "inpoint", (guint64) 0, "priority", 1, NULL);
+      "duration", (gint64) 10 * GST_SECOND, "inpoint", (guint64) 0, "priority",
+      1, NULL);
 
   nle_composition_add (composition, source2);
 
@@ -317,7 +319,8 @@ GST_START_TEST (test_dispose_on_commit)
   audiotestsrc = gst_element_factory_make ("audiotestsrc", "audiotestsrc1");
   gst_bin_add (GST_BIN (nlesource), audiotestsrc);
   g_object_set (nlesource, "start", (guint64) 0 * GST_SECOND,
-      "duration", 10 * GST_SECOND, "inpoint", (guint64) 0, "priority", 1, NULL);
+      "duration", (gint64) 10 * GST_SECOND, "inpoint", (guint64) 0, "priority",
+      1, NULL);
   fail_unless (nle_composition_add (GST_BIN (composition), nlesource));
 
   gst_bin_add_many (GST_BIN (pipeline), composition, fakesink, NULL);
@@ -373,7 +376,8 @@ GST_START_TEST (test_simple_audiomixer)
   audiotestsrc1 = gst_element_factory_make ("audiotestsrc", "audiotestsrc1");
   gst_bin_add (GST_BIN (nlesource1), audiotestsrc1);
   g_object_set (nlesource1, "start", (guint64) 0 * GST_SECOND,
-      "duration", total_time / 2, "inpoint", (guint64) 0, "priority", 1, NULL);
+      "duration", (gint64) total_time / 2, "inpoint", (guint64) 0, "priority",
+      1, NULL);
   fail_unless (nle_composition_add (GST_BIN (composition), nlesource1));
 
   /* nlesource2 */
@@ -463,7 +467,8 @@ create_nested_source (gint nesting_depth)
     name = g_strdup_printf ("nested_src%d", i);
     source = gst_element_factory_make_or_warn ("nlesource", name);
     g_free (name);
-    g_object_set (source, "start", 0, "duration", 2 * GST_SECOND, NULL);
+    g_object_set (source, "start", (guint64) 0, "duration",
+        (gint64) 2 * GST_SECOND, NULL);
     gst_bin_add (GST_BIN (source), bin);
   }
